+++ backend/club4u/test/__init__.py
@@ -1,0 +1,5 @@
+from .test_general import *
+from .test_user import *
+from .test_club import *
+from .test_somoim import *
+from .test_preclub import *
